requirements.txt
Flask==2.3.2
google-play-scraper==1.2.4
pandas==2.0.3
sastrawi==1.0.1
nltk==3.8.1
wordcloud==1.9.2
matplotlib==3.7.2
python-dotenv==1.0.0